feat: switch to uint16_t

+++ include/beeb-boop-beep.h
@@ -1,6 +1,8 @@
 // https://dopeloop.ai/melody-generator
 
-int bbboteFrequencies[] = {349, 523, 440};
-int noteDurations[] = {62, 62, 62};
-int notePauses[] = {62, 62, 187};
-int numberOfNotes = 3;
+#include <Arduino.h>
+
+uint16_t bbboteFrequencies[] = {349, 523, 440};
+uint16_t noteDurations[] = {62, 62, 62};
+uint16_t notePauses[] = {62, 62, 187};
+uint16_t numberOfNotes = 3;
